Admin panel security, login form and admins
The administrator panel is now secured behind a login form. Login forms uses the same LoginManager as the REST API and users can be marked as "is_admin" (boolean) which determines whether they can access the admin panel or not. The admin panel allows for easy database data insertion, deletion or modification.